# Handin 3/Backend/docs/swagger.yaml
definitions:
  models.Measurement:
    properties:
      createdDate:
        type: string
      id:
        type: integer
      measurementType:
        type: string
      value:
        type: integer
    type: object
  models.StatusDTO:
    properties:
      alive:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v1/mqtt-connection:
    get:
      consumes:
      - application/json
      description: Indicates if api is connected to MQTT broker
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusDTO'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Indicates if api is connected to MQTT broker
      tags:
      - GET API Endpoints
  /api/v1/sensors/air-quality/co2:
    get:
      consumes:
      - application/json
      description: Gets the latest Co2 measurement from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Measurement'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get latest Co2 measurement
      tags:
      - GET API Endpoints
  /api/v1/sensors/air-quality/tvoc:
    get:
      consumes:
      - application/json
      description: Gets the latest VOC measurement from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Measurement'
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get latest TVOC measurement
      tags:
      - GET API Endpoints
  /led-state/{on}:
    put:
      consumes:
      - application/json
      description: Change led state on IoT device
      parameters:
      - description: State of led, 0 = off, 1 = on
        in: path
        name: "on"
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Change led state on IoT device
      tags:
      - PUT API Endpoints
swagger: "2.0"